Introduce support to build a subset of tests
The build system now supports building a TFTF binary with a subset of
tests. The chosen subset may be specified through the TESTS variable,
e.g.:

 make TESTS=sdei tftf
 make TESTS=tsp tftf
 make TESTS=cpu-extensions tftf

This will use tftf/tests/tests-${TESTS}.xml as the tests manifest
          and tftf/tests/tests-${TESTS}.mk as the tests makefile.

This provides more flexibility than before. It is still possible to
compile all the tests using:

 make TESTS=common tftf

It's also possible to include the extensive tests using:

 make TESTS=extensive tftf

If TESTS is not specified, it defaults to the 'common' tests.

Change-Id: Ifcdc52e9a29ed5f9a9bf57703cdb109b5a169fa2
Signed-off-by: Sandrine Bailleux <sandrine.bailleux@arm.com>