Do not print CPU MPID in mp_printf()
mp_printf() should just be an MP-safe version of printf(), i.e. one
that takes the console lock before printing. It should not be
responsible for printing the CPU MPID as well, this decision should be
left to the caller.

Also make Cactus and Ivy use mp_printf(). Before that, they could not
call this function because they couldn't access the MPIDR_EL1 as
S-EL0 images.

Change-Id: I4eafee01ffc279296395b94dd4a07cfbb8e858e2
Signed-off-by: Sandrine Bailleux <sandrine.bailleux@arm.com>