stdlib: Increase compatibility with TF-A
Code that originates from the Trusted Firmware-A project expects the
stdlib headers to work in a specific way and to have some specific
defines. Specifically, TF-A doesn't have the non-standard types.h
header, and it has all the definitions in stdint.h.

Also, the __init define is missing. No component of this repository
needs this option, but having the define in cdefs.h allows code sharing
between both projects.

Change-Id: Ic298fd87a6c2cf8a9e5b8a18fc274d4150ed0a13
Signed-off-by: Antonio Nino Diaz <antonio.ninodiaz@arm.com>