Remove MPIDR_{CLUSTER,CPU}_ID macros from arch.h
Hardcoding CPU to affinity level 0 and cluster to level 1 isn't correct.
This patch removes the definitions from arch.h to prevent more tests from
making this assumption. It doesn't fix the tests that are already using
them as it may be needed to do more changes to make the tests truly
generic.

Change-Id: I3e5362ef7db7769f7db98ba68ad3842f5baa3e60
Signed-off-by: Antonio Nino Diaz <antonio.ninodiaz@arm.com>