Fix waitus()
waitus() was missing an instruction barrier before reading the system
counter value. As a result, the CPU could speculatively read it and
we would potentially base our delay loop on stale values.

waitus() now uses the syscounter_read() helper function introduced in
the previous patch, which has the required ISB.

Change-Id: Ic37254485a9cdc4d4d2c86d245aa3273454e82ff
Signed-off-by: Sandrine Bailleux <sandrine.bailleux@arm.com>