Relax SMC tests when Trusted OS is present
When receiving an invalid SMC function ID, the Trusted OS is free to
return any error code in x0 but it should at least preserve the values
of x1-x3. The SMC tests expected the Trusted OS to return -1 in x0,
which is too strong of an assumption. This patch relaxes them to check
x1-x3 but ignore x0.

Change-Id: I6d516319afeaee9af6aa3674a48c13dba6f7af69
Signed-off-by: Sandrine Bailleux <sandrine.bailleux@arm.com>