Move unstable tests to a separate pool of tests
Some tests have proven unreliable over time. They often time out and the
root cause of the problem is unknown as of today.

This root cause needs to be investigated and fixed but until then, this
patch isolates these tests in their own pool. This allows us to set
different trust levels in our CI:
 - The stable tests are trusted to call out the validity of a patch.
 - The unstable tests are only informative.

Change-Id: Ia8ebb5a29a523ef3a1f324dbd3ddddac02254fd8
Signed-off-by: Sandrine Bailleux <sandrine.bailleux@arm.com>