Rework IRQ vector code
Jump straight to a function so that we are not constrained in the
32-instruction limit. Turn the code saving and restoring
general-purpose registers into a macro (as opposed to a function),
thus removing the need to save the link register before anything else.

Overall, this makes the code more straight-forward as we can more
clearly see that registers are saved from x0 to x30 (in this order)
without any magic twist.

Change-Id: Ic70ceee62bec5efc6f6b6a23a7bd4c1d6c2f8a97
Signed-off-by: Sandrine Bailleux <sandrine.bailleux@arm.com>