Dump some registers when hitting an unexpected exception
At the moment, no information is printed on the UART whenever we hit
an unexpected exception, not even an error message. This is not great
from the user's perspective, who has got no idea of what is going on.

Now we print an error message, as well as the state of some of the
registers. This includes general-purpose registers, as well as some
system registers.

This is implemented for TFTF running:

- in AArch64 state, at EL2;
- in AArch64 state, at NS-EL1;
- in AArch32 state.

We might want to dump more registers in the future but this patch
at least provides a basis we can build upon.

Also, the SP_EL0 has been removed from the list of registers saved in
the CPU context because TFTF always uses SP_ELx and does not touch
SP_EL0 at all.

Change-Id: I56e4afa917b53b5ccccff1d5d09ac8ccfaa6ae49
Signed-off-by: Sandrine Bailleux <sandrine.bailleux@arm.com>