Remove unexecuted code for AMU group 1 counters
This patch removes from the AMU tests the part that iterates over group
1 counters. The reasons for this are the following:

1) Currently the AMU tests are executed only in FVP but FVP doesn't
have/expose group 1 counters.

2) Even if some platform implements group 1 counters in the future, we
don't know what they will be counting. As such we cannot know what the
valid values for those counters will be. Hence, we can't establish a
specific test at this moment to validate the group 1 counter values.

3) Even now that the loop exists in the code, the macro defining the
group 1 counters is defined as zero. Which means that the loop is
already not executed, despite being there (dead code).

However, the architecture defines how group 1 counters should be
implemented. For this reason, the assembly helper functions that access
the group1 counters are left intact. This will allow us to easily extend
the AMU tests again in the future to include group 1 counters.

Change-Id: I5a6f119c7e817de9dc5e510c3fadd008820bd564
Signed-off-by: John Tsichritzis <john.tsichritzis@arm.com>