Bugfix for PMU leakage test
This patch fixes a bug where integer underflow causes the tests
to wrongly fail.

Since event counts are register values, they are unsigned in TFTF.
When the event count is less on the SMC being profiled than the
baseline SMC event count, subtraction causes an underflow which
goes beyond the allowed deviation and makes the test fail.

Signed-off-by: Petre-Ionut Tudor <petre-ionut.tudor@arm.com>
Change-Id: I58bc18ca4afd28b6d1b1354a9af9f70d616d2c32