Aligning block_buffer to uint32_t
block_buffer will eventually be used by the flash_write_buffer() implementation
of the NOR driver for the Juno board. That implementation requires the buffer
passed to it to be 32bit aligned, which might not be enforced depending on
the compiler optimizations.
This patch will enforce that alignment on all cases.

Change-Id: I84d22f6b796d5d6c93d3c69dd24c7ebe01103691
Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>