Trim down the top-level readme file
The main TF-A Tests documentation on ReadTheDocs does not use the
top-level readme file. However, as explained in commit f6ee8fbcef6e
("Revert "Remove unused top-level readme file""), we still want to
keep it because it is the page displayed on
https://git.trustedfirmware.org/TF-A/tf-a-tests.git/about.

That said, this readme file does not need to provide such detailed
information, instead it can simply be a landing page providing a brief
overview of the project and redirecting the reader to RTD for further
information.

Revisit the top-level readme file in this way, and also fix/update a
number of broken/out-dated links in it.

Change-Id: If53cfa9bd0bf2cba044883bb7ba5e5114bfb2bf8
Signed-off-by: Sandrine Bailleux <sandrine.bailleux@arm.com>