fix(cactus): remove misleading ERROR log
The cactus message loop the log message for unhandled commands is
misleading because it is expected to receive some direct messages for
set-up configuration discovery. These messages are not handled by cactus
and could indicate when a given test must be skipped or not. Namely,
some SPM tests are skipped if OPTEE deployed as SP. This discovery
"protocol" is executed at the beginning of most SPM related tests, which
lead to a lot of ERROR messages, that often mislead developers into
thinking there was a problem in the system.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: Ib14dc3bf7f9c916fd5addf24675ea111eef5248e