test(notifications): NPI for per-vCPU notifications
Tested that NPI is injected into and handled by the expected vCPU in a
MP setup.
- Added functions to bring up all cores, and enable/disable the NPI
vIRQ, in the SP's vIRQ.
- Refactored previously implemented test to per-vCPU notifications,
such that (similarly to global notifications test), it is verified that
the NPI is handled in each one of the cores, upon requesting cactus SP
to get its pending notifications.

Change-Id: Ib1d003aa86538956c94577578fd6ae9a08cf44b0
Signed-off-by: J-Alves <joao.alves@arm.com>