fix(build): use common name for ivy partitions
Trusted firmware sp_gen.mk expects the package name to be ivy.pkg
for all ivy packages (both with and without the shim). However
the packages are named according to the partition name in
sp_layout.json. Make ivy partitions use the ivy name in
sp_layout.json to prevent this name mismatch.

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: If4702667b13be26feee809d10b21c46110ff6848