arm64: Use the actual memory size of the images to place images
The arm64 kernel image header (also used by EDK2 UEFI) provides the
actual size of the kernel image in memory, including the bss and any
other reserved pages for early boot. This is not accounted for in the
file size, and we could potentially place other images at this
"uninitialized" area of the kernel (e.g., initrd), which may get
overwritten during the early boot. Use the actual memory size of the
image, when available, to place the images safely.

Signed-off-by: Suzuki K Poulose <suzuki.poulose@arm.com>