Add handling of the realm metadata
From now on, the kvmtool can use an additional command
line option --metadata <filename> to provide the path
to the realm metadata file containing identity and authority
information of the realm.

Once the option is provided, the realm metadata is passed
to the kvm layer and then to the RMM.

The realm metadata can be utilized by the RMM's sealing
key derivation mechanism to retrieve keys resistant to
realm updates.

Signed-off-by: Piotr Sawicki <p.sawicki2@samsung.com>