Add test_dma_ops to virtio_pci_modern
If the dma_ops is null, vring_alloc_queue() calls dma_direct_alloc()
But if the dma_ops is not null, then it calls dma_ops->alloc().

Note: when you use dma_ops, there are essential non-null callbacks.
Please see the test_dma_ops in the virtio_pci_modern.c for the callbacks

Signed-off-by: Sunwook Eom <speed.eom@samsung.com>