Add support for running QEMU with external software HES
This adds a new define TF_A_RSE_SERIAL_HES that causes:
- MEASURED_BOOT=1, PLAT_RSE_COMMS_USE_SERIAL=1 and mbedtls compilation in TF-A
- running emulated software TPM (swtpm) that is required now
- adding TPM device in QEMU so it can actually use it
- reconfiguring second serial port as it's required for the HES to work

The configuration could be simplified if we didn't care about the terminal on
the second serial port in case we run without HES, but I don't know if it's not
utilized by something else.

Without the TF_A_RSE_SERIAL_HES set the behaviour should be unchanged.

Signed-off-by: Lukasz Pawelczyk <l.pawelczyk@samsung.com>